+++ src/LDL/Framework/Base/Collection/CallableCollectionInterface.php
@@ -1,0 +1,20 @@
+<?php declare(strict_types=1);
+
+namespace LDL\Framework\Base\Collection;
+
+use LDL\Framework\Base\Collection\Contracts\AppendableInterface;
+use LDL\Framework\Base\Collection\Contracts\CollectionInterface;
+use LDL\Framework\Base\Collection\Contracts\LockAppendInterface;
+use LDL\Framework\Base\Collection\Contracts\LockRemoveInterface;
+use LDL\Framework\Base\Collection\Contracts\RemovableInterface;
+use LDL\Framework\Base\Contracts\LockableObjectInterface;
+
+interface CallableCollectionInterface extends CollectionInterface, AppendableInterface, LockAppendInterface, RemovableInterface, LockRemoveInterface, LockableObjectInterface
+{
+
+    /**
+     * @param mixed ...$params
+     */
+    public function call(...$params) : void;
+
+}