+++ src/LDL/Framework/Base/Collection/Contracts/ReplaceableInterface.php
@@ -9,12 +9,6 @@ use LDL\Framework\Base\Exception\LockingException;
 interface ReplaceableInterface
 {
     /**
-     * @param $item
-     * @param $key
-     */
-    public function onBeforeReplace($item, $key): void;
-
-    /**
      * If the key already exists, it will be replaced, if the key does not exists
      * it will throw a ReplaceException
      *