+++ src/LDL/Framework/Base/Collection/Traits/TruncateInterfaceTrait.php
@@ -6,6 +6,7 @@
 
 namespace LDL\Framework\Base\Collection\Traits;
 
+use LDL\Framework\Base\Collection\Contracts\BeforeRemoveInterface;
 use LDL\Framework\Base\Collection\Contracts\CollectionInterface;
 use LDL\Framework\Base\Collection\Contracts\LockRemoveInterface;
 use LDL\Framework\Base\Contracts\LockableObjectInterface;
@@ -22,8 +23,12 @@ trait TruncateInterfaceTrait
         if($this instanceof LockRemoveInterface){
             $this->checkLockRemove();
         }
-        foreach($this as $key => $item) {
-            $this->onBeforeRemove($item, $key);
+
+        if($this instanceof BeforeRemoveInterface) {
+            $before = $this->getBeforeRemove();
+            foreach ($this as $key => $item) {
+                $before->call($this, $item, $key);
+            }
         }
 
         $collection = clone $this;